+++ src/test/java/ru/praktikum/yandex/RegisterFieldsValidationTest.java
@@ -24,14 +24,10 @@ public class RegisterFieldsValidationTest extends TestBase {
     UserClient userClient = new UserClient();
 
     @After
-    public void teardown() {
+    public void tearDown() {
         //удаляем пользователя после теста
-        try {
+        if (user.getAccessToken() != null){
             userClient.deleteUser(user);
-            System.out.println("User deleted after Test");
-        } catch (NullPointerException e) {
-            user = null;
-
         }
         driver.quit();
     }
